jsalamero/installer improvements

+++ extra/installer/build_cd.sh
@@ -24,11 +24,14 @@ pushd $TMPDIR/zenbuntu-desktop
 dpkg-buildpackage
 popd
 
-# Build zinstaller-remote udeb
-cp -rL zinstaller-remote $TMPDIR/zinstaller-remote
-pushd $TMPDIR/zinstaller-remote
-dpkg-buildpackage
-popd
+if [ "$INCLUDE_REMOTE" == "true" ]
+then
+    # Build zinstaller-remote udeb
+    cp -rL zinstaller-remote $TMPDIR/zinstaller-remote
+    pushd $TMPDIR/zinstaller-remote
+    dpkg-buildpackage
+    popd
+fi
 
 for ARCH in $ARCHS
 do
@@ -48,11 +51,14 @@ do
     rm $EXTRAS_DIR/zenbuntu-desktop_*.deb
     cp $TMPDIR/*.deb $EXTRAS_DIR/
 
-    # Add zinstaller-remote udeb
-    UDEB_DIR=$CD_BUILD_DIR/pool/main/z/zinstaller-remote
-    mkdir -p $UDEB_DIR
-    rm $UDEB_DIR/*
-    cp $TMPDIR/*.udeb $UDEB_DIR/
+    if [ "$INCLUDE_REMOTE" == "true" ]
+    then
+        # Add zinstaller-remote udeb
+        UDEB_DIR=$CD_BUILD_DIR/pool/main/z/zinstaller-remote
+        mkdir -p $UDEB_DIR
+        rm $UDEB_DIR/*
+        cp $TMPDIR/*.udeb $UDEB_DIR/
+    fi
 
     test -d $CD_BUILD_DIR/isolinux || (echo "isolinux directory not found in $CD_BUILD_DIR."; false) || exit 1
     test -d $CD_BUILD_DIR/.disk || (echo ".disk directory not found in $CD_BUILD_DIR."; false) || exit 1
